Marketing Recommendations for Teramind
The following are marketing recommendations with assumed figures that are not meant to be shared without express permission from Ron Kagan. The recommendations are informed by the bowtie revenue model from Winning by Design and observed opportunities for growth. This is a presentation that is not meant to be used for commercial purposes.